Fixed some small issues
- Issue where game would not return result if the computer player made the
  move and then that resulted in checkmate/stalemate
- Tweak to filehandling for clearing a file
- Changed the heuristic calculation to be a difference of the two
  functions that is returned

+++ filehandling.py
@@ -4,7 +4,7 @@ import os
 def check4Move(player, moveNum):
 	#MoveNum must start at 1 and go up to N
 	moveList = []
-	duration = 2
+	duration = 1
 	if player == "X":
 		file = "log_y.txt"
 	if player == "Y":
@@ -27,9 +27,10 @@ def check4Move(player, moveNum):
 	return moveList[-1] #Return last element of the list
 	
 def emptyLogFiles():
-	with open("log_x.txt", "w") as f:
-		f.seek(0)
-		f.truncate()
-	with open("log_y.txt", "w") as f:
-		f.seek(0)
-		f.truncate()
+	if os.path.getsize("log_x.txt") >= 7:
+		with open("log_x.txt", "w") as f:
+			f.seek(0)
+			f.truncate()
+		with open("log_y.txt", "w") as f:
+			f.seek(0)
+			f.truncate()
